Initial commit to refactor message parsing into separate module.
